change Configuration to u16
Due to kernel source changes, Configuration variable must be u16
instead of ULONG. The AND operation that removes value of last 2 bits
is modified to match size of u16.

le2cpu16 is removed as it does not support u16 type. I am not so sure if
this should be removed. // TODO: investigate

Signed-off-by: tomsik68 <tomsik68@gmail.com>